+++ CHANGES.md
@@ -1,6 +1,9 @@
 2.0.1 (unreleased)
 =====
 
+New:
+* Added `time.predicate` to parse time predicates at runtime.
+
 Changed:
 * Removed `encode_metadata` option in `input.file.hls` as it does nothing with
   the main encoder for HLS format, `%ffmpeg` (#2023)